Register panic button uuid when entering module
